discord_bot/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.txt -o requirements.txt
aiofiles==25.1.0
    # via
    #   -r requirements.txt
    #   quart
aiohappyeyeballs==2.6.1
    # via
    #   -r requirements.txt
    #   aiohttp
aiohttp==3.13.2
    # via
    #   -r requirements.txt
    #   discord-py
aiomysql==0.3.2
    # via -r requirements.txt
aiosignal==1.4.0
    # via
    #   -r requirements.txt
    #   aiohttp
anyio==4.12.1
    # via httpx
attrs==25.4.0
    # via
    #   -r requirements.txt
    #   aiohttp
blinker==1.9.0
    # via
    #   -r requirements.txt
    #   flask
    #   quart
cachetools==6.2.4
    # via
    #   -r requirements.txt
    #   flask-discord
certifi==2025.11.12
    # via
    #   -r requirements.txt
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via
    #   -r requirements.txt
    #   requests
click==8.3.1
    # via
    #   -r requirements.txt
    #   flask
    #   quart
discord-py==2.6.4
    # via -r requirements.txt
flask==3.1.2
    # via
    #   -r requirements.txt
    #   flask-discord
    #   quart
flask-discord==0.1.69
    # via -r requirements.txt
frozenlist==1.8.0
    # via
    #   -r requirements.txt
    #   aiohttp
    #   aiosignal
greenlet==3.3.0
    # via
    #   -r requirements.txt
    #   sqlalchemy
h11==0.16.0
    # via
    #   -r requirements.txt
    #   httpcore
    #   hypercorn
    #   wsproto
h2==4.3.0
    # via
    #   -r requirements.txt
    #   hypercorn
hpack==4.1.0
    # via
    #   -r requirements.txt
    #   h2
httpcore==1.0.9
    # via httpx
httpx==0.27.0
    # via -r requirements.txt
hypercorn==0.18.0
    # via
    #   -r requirements.txt
    #   quart
hyperframe==6.1.0
    # via
    #   -r requirements.txt
    #   h2
idna==3.11
    # via
    #   -r requirements.txt
    #   anyio
    #   httpx
    #   requests
    #   yarl
itsdangerous==2.2.0
    # via
    #   -r requirements.txt
    #   flask
    #   quart
jinja2==3.1.6
    # via
    #   -r requirements.txt
    #   flask
    #   quart
markupsafe==3.0.3
    # via
    #   -r requirements.txt
    #   flask
    #   jinja2
    #   quart
    #   werkzeug
multidict==6.7.0
    # via
    #   -r requirements.txt
    #   aiohttp
    #   yarl
mysql-connector-python==9.5.0
    # via -r requirements.txt
numpy==2.3.5
    # via
    #   -r requirements.txt
    #   pandas
oauthlib==3.3.1
    # via
    #   -r requirements.txt
    #   flask-discord
    #   requests-oauthlib
pandas==2.3.3
    # via -r requirements.txt
priority==2.0.0
    # via
    #   -r requirements.txt
    #   hypercorn
propcache==0.4.1
    # via
    #   -r requirements.txt
    #   aiohttp
    #   yarl
pyjwt==2.10.1
    # via
    #   -r requirements.txt
    #   flask-discord
pymysql==1.1.2
    # via
    #   -r requirements.txt
    #   aiomysql
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.txt
    #   pandas
python-dotenv==1.2.1
    # via -r requirements.txt
pytz==2025.2
    # via
    #   -r requirements.txt
    #   pandas
quart==0.20.0
    # via
    #   -r requirements.txt
    #   quart-cors
quart-cors==0.8.0
    # via -r requirements.txt
requests==2.32.5
    # via
    #   -r requirements.txt
    #   flask-discord
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via
    #   -r requirements.txt
    #   flask-discord
six==1.17.0
    # via
    #   -r requirements.txt
    #   python-dateutil
sniffio==1.3.1
    # via httpx
sqlalchemy==2.0.45
    # via -r requirements.txt
typing-extensions==4.15.0
    # via
    #   -r requirements.txt
    #   aiosignal
    #   anyio
    #   sqlalchemy
tzdata==2025.3
    # via
    #   -r requirements.txt
    #   pandas
urllib3==2.6.2
    # via
    #   -r requirements.txt
    #   requests
werkzeug==3.1.4
    # via
    #   -r requirements.txt
    #   flask
    #   quart
wsproto==1.3.2
    # via
    #   -r requirements.txt
    #   hypercorn
yarl==1.22.0
    # via
    #   -r requirements.txt
    #   aiohttp
